requirements.txt
flask
requests
Pillow
werkzeug
gunicorn


